refactor(gateway): split server implementation
