+++ src/gateway/server-startup-bootstrap.ts
@@ -1,0 +1,558 @@
+import { getActiveBackgroundExecSessionCount } from "../agents/bash-process-registry.js";
+import { getActiveEmbeddedRunCount } from "../agents/embedded-agent-runner/run-state.js";
+import { getTotalPendingReplies } from "../auto-reply/reply/dispatcher-registry.js";
+import { isRestartEnabled } from "../config/commands.flags.js";
+import {
+  collectConfigRuntimeEnvOwnership,
+  initializePublishedConfigRuntimeEnv,
+  prepareConfigRuntimeEnv,
+} from "../config/config-env-vars.js";
+import { assertGatewayConfigEnvSelectionUnchanged } from "../config/gateway-env-selection.js";
+import {
+  getRuntimeConfig,
+  getRuntimeConfigSourceSnapshot,
+  readConfigFileSnapshot,
+  setAppliedRuntimeConfigSnapshot,
+} from "../config/io.js";
+import { normalizeStateDirEnv } from "../config/paths.js";
+import { captureConfigOverrideApplier } from "../config/runtime-overrides.js";
+import { resolveMainSessionKey } from "../config/sessions.js";
+import type { GatewayAuthConfig } from "../config/types.gateway.js";
+import type { OpenClawConfig } from "../config/types.openclaw.js";
+import { isSecretRef } from "../config/types.secrets.js";
+import { getActiveCronJobCount } from "../cron/active-jobs.js";
+import {
+  listDevicePairing,
+  resolveEffectiveOperatorDeviceIdentity,
+  type EffectiveOperatorDeviceIdentity,
+} from "../infra/device-pairing.js";
+import {
+  isDiagnosticsEnabled,
+  setDiagnosticsEnabledForProcess,
+} from "../infra/diagnostic-events.js";
+import { isVitestRuntimeEnv, logAcceptedEnvOption } from "../infra/env.js";
+import { readGatewayRestartHandoffSync } from "../infra/restart-handoff.js";
+import { setGatewaySigusr1RestartPolicy, setPreRestartDeferralCheck } from "../infra/restart.js";
+import { enqueueSystemEvent } from "../infra/system-events.js";
+import { startDiagnosticHeartbeat } from "../logging/diagnostic.js";
+import type { createSubsystemLogger } from "../logging/subsystem.js";
+import { setCurrentPluginMetadataSnapshot } from "../plugins/current-plugin-metadata-snapshot.js";
+import { getTotalQueueSize } from "../process/command-queue.js";
+import { getActiveGatewayRootWorkCount } from "../process/gateway-work-admission.js";
+import { createLazyPromise } from "../shared/lazy-runtime.js";
+import { roleScopesAllow } from "../shared/operator-scope-compat.js";
+import { ADMIN_SCOPE } from "./method-scopes.js";
+import { listCoreGatewayMethodNames } from "./methods/core-descriptors.js";
+import {
+  mergeActivationSectionsIntoRuntimeConfig,
+  resolveGatewayReloadPluginActivationCandidate,
+} from "./plugin-activation-runtime-config.js";
+import {
+  resumeGatewayRestartTraceFromEnv,
+  resumeGatewayRestartTraceFromHandoff,
+} from "./restart-trace.js";
+import type { GatewayServerOptions } from "./server-public.js";
+import { createGatewayStartupTrace } from "./server-startup-trace.js";
+import { mergeGatewayAuthConfig, mergeGatewayTailscaleConfig } from "./startup-auth.js";
+import { maybeSeedControlUiAllowedOriginsAtStartup } from "./startup-control-ui-origins.js";
+
+type GatewayLogger = ReturnType<typeof createSubsystemLogger>;
+type WorkerEnvironmentStartupLoader = () => Promise<
+  typeof import("./server-worker-environment-startup.js")
+>;
+
+export async function prepareGatewayServerBootstrap(input: {
+  port: number;
+  opts: GatewayServerOptions;
+  log: GatewayLogger;
+  logSecrets: GatewayLogger;
+  loadWorkerEnvironmentStartupModule: WorkerEnvironmentStartupLoader;
+  formatRuntimeGatewayAuthTokenWarning: () => string;
+}) {
+  const { port, opts, log, logSecrets, loadWorkerEnvironmentStartupModule } = input;
+  const formatRuntimeGatewayAuthTokenWarning = input.formatRuntimeGatewayAuthTokenWarning;
+  normalizeStateDirEnv(process.env);
+  const [
+    {
+      OPENCLAW_DATABASE_SCHEMA_DOCS_URL,
+      OpenClawDatabaseSchemaPreflightError,
+      preflightOpenClawDatabaseSchemas,
+    },
+    agentDatabase,
+    stateDatabase,
+  ] = await Promise.all([
+    import("../state/openclaw-database-preflight.js"),
+    import("../state/openclaw-agent-db.js"),
+    import("../state/openclaw-state-db.js"),
+  ]);
+  const databaseSchemas = preflightOpenClawDatabaseSchemas({
+    env: process.env,
+    supportedVersions: {
+      state: stateDatabase.OPENCLAW_STATE_SCHEMA_VERSION,
+      agent: agentDatabase.OPENCLAW_AGENT_SCHEMA_VERSION,
+    },
+  });
+  if (databaseSchemas.incompatible.length > 0) {
+    for (const database of databaseSchemas.incompatible) {
+      log.error("database schema preflight rejected newer schema", {
+        kind: database.kind,
+        path: database.path,
+        ...(database.agentId ? { agentId: database.agentId } : {}),
+        foundVersion: database.foundVersion,
+        supportedVersion: database.supportedVersion,
+        writerAppVersion: database.writerAppVersion ?? "unknown",
+        docsUrl: OPENCLAW_DATABASE_SCHEMA_DOCS_URL,
+      });
+    }
+    throw new OpenClawDatabaseSchemaPreflightError(databaseSchemas.incompatible);
+  }
+  for (const database of databaseSchemas.indeterminate) {
+    log.warn("database schema preflight could not inspect database; continuing to real open", {
+      kind: database.kind,
+      path: database.path,
+      reason: database.reason,
+      docsUrl: OPENCLAW_DATABASE_SCHEMA_DOCS_URL,
+    });
+  }
+  const { bootstrapGatewayNetworkRuntime } = await import("./server-network-runtime.js");
+  bootstrapGatewayNetworkRuntime();
+
+  const minimalTestGateway =
+    isVitestRuntimeEnv() && process.env.OPENCLAW_TEST_MINIMAL_GATEWAY === "1";
+  const ambientEnvTriggers = opts.ambientEnvTriggers ?? "allow";
+
+  // Ensure all default port derivations (browser/canvas) see the actual runtime port.
+  process.env.OPENCLAW_GATEWAY_PORT = String(port);
+  logAcceptedEnvOption({
+    key: "OPENCLAW_RAW_STREAM",
+    description: "raw stream logging enabled",
+  });
+  logAcceptedEnvOption({
+    key: "OPENCLAW_RAW_STREAM_PATH",
+    description: "raw stream log path override",
+  });
+  if (!resumeGatewayRestartTraceFromEnv(process.env, [["source", "env"]])) {
+    const restartHandoff = readGatewayRestartHandoffSync();
+    resumeGatewayRestartTraceFromHandoff(restartHandoff?.restartTrace, [
+      ["source", restartHandoff?.source],
+      ["restartKind", restartHandoff?.restartKind],
+      ["supervisorMode", restartHandoff?.supervisorMode],
+    ]);
+  }
+  const startupTrace = createGatewayStartupTrace(log);
+  const startupConfigModulePromise = import("./server-startup-config.js");
+  const loadStartupPluginsModule = createLazyPromise(() => import("./server-startup-plugins.js"), {
+    cacheRejections: true,
+  });
+  const { loadGatewayStartupConfigSnapshot } = await startupConfigModulePromise;
+
+  const envBeforeStartupConfigLoad = { ...process.env };
+  const startupConfigLoad = await startupTrace.measure("config.snapshot", () =>
+    loadGatewayStartupConfigSnapshot({
+      minimalTestGateway,
+      log,
+      measure: (name, run) => startupTrace.measure(name, run),
+      ...(opts.startupConfigSnapshotRead
+        ? { initialSnapshotRead: opts.startupConfigSnapshotRead }
+        : {}),
+    }),
+  );
+  const configSnapshot = startupConfigLoad.snapshot;
+  const startupAuthOverride = opts.auth ? structuredClone(opts.auth) : undefined;
+  const startupTailscaleOverride = opts.tailscale ? structuredClone(opts.tailscale) : undefined;
+  // Seed before secrets activation so every active/rollback snapshot carries
+  // the same runtime-only browser origin baseline.
+  const controlUiSeed = minimalTestGateway
+    ? { config: configSnapshot.config, seededAllowedOrigins: false }
+    : await startupTrace.measure("control-ui.seed", () =>
+        maybeSeedControlUiAllowedOriginsAtStartup({
+          config: configSnapshot.config,
+          log,
+          runtimeBind: opts.bind,
+          runtimePort: port,
+        }),
+      );
+  const startupConfigSnapshot = controlUiSeed.seededAllowedOrigins
+    ? {
+        ...configSnapshot,
+        runtimeConfig: controlUiSeed.config,
+        config: controlUiSeed.config,
+      }
+    : configSnapshot;
+
+  const emitSecretsStateEvent = (
+    code: "SECRETS_RELOADER_DEGRADED" | "SECRETS_RELOADER_RECOVERED",
+    message: string,
+    cfg: OpenClawConfig,
+  ) => {
+    enqueueSystemEvent(`[${code}] ${message}`, {
+      sessionKey: resolveMainSessionKey(cfg),
+      contextKey: code,
+    });
+  };
+  const { createRuntimeSecretsActivator } = await startupConfigModulePromise;
+  const activateRuntimeSecrets = createRuntimeSecretsActivator({
+    logSecrets,
+    emitStateEvent: emitSecretsStateEvent,
+    channelAutostartSuppression: opts.channelAutostartSuppression,
+    ...(startupConfigLoad.pluginMetadataSnapshot
+      ? { pluginMetadataSnapshot: startupConfigLoad.pluginMetadataSnapshot }
+      : {}),
+  });
+  let startupInternalWriteHash: string | null = null;
+  let startupLastGoodSnapshot = configSnapshot;
+  const startupActivationSourceConfig = configSnapshot.sourceConfig;
+  const startupRuntimeConfig = captureConfigOverrideApplier()(startupConfigSnapshot.config);
+  startupTrace.setConfig(startupRuntimeConfig);
+  const { prepareGatewayStartupConfig } = await startupConfigModulePromise;
+  const authBootstrap = await startupTrace.measure(
+    "config.auth",
+    () =>
+      prepareGatewayStartupConfig({
+        configSnapshot: startupConfigSnapshot,
+        authOverride: startupAuthOverride,
+        tailscaleOverride: startupTailscaleOverride,
+        activateRuntimeSecrets,
+        log,
+        measure: (name, run, measureOptions) => startupTrace.measure(name, run, measureOptions),
+      }),
+    { omitErrorMessage: true },
+  );
+  const cfgAtStart = authBootstrap.cfg;
+  startupTrace.setConfig(cfgAtStart);
+  const {
+    claimControlUiDeviceAuthMigration,
+    completeControlUiDeviceAuthMigration,
+    importPendingControlUiDeviceAuthMigration,
+    isLegacyControlUiDeviceAuthMigrationInput,
+    readControlUiDeviceAuthMigrationState,
+    recoverControlUiDeviceAuthMigrationClaim,
+    releaseControlUiDeviceAuthMigrationClaim,
+  } = await import("../state/control-ui-device-auth-migration.js");
+  const legacyControlUiDeviceAuthBypass = isLegacyControlUiDeviceAuthMigrationInput({
+    disabledDeviceAuth: cfgAtStart.gateway?.controlUi?.dangerouslyDisableDeviceAuth === true,
+    lastTouchedVersion: cfgAtStart.meta?.lastTouchedVersion,
+  });
+  let controlUiDeviceAuthMigrationState = legacyControlUiDeviceAuthBypass
+    ? importPendingControlUiDeviceAuthMigration({ env: process.env })
+    : readControlUiDeviceAuthMigrationState({ env: process.env });
+  if (
+    controlUiDeviceAuthMigrationState?.status === "pending" &&
+    controlUiDeviceAuthMigrationState.claimedDeviceId
+  ) {
+    // A process crash between claim and approval must not strand the upgrade.
+    controlUiDeviceAuthMigrationState = recoverControlUiDeviceAuthMigrationClaim({
+      env: process.env,
+    });
+  }
+  if (controlUiDeviceAuthMigrationState?.status === "pending") {
+    const existingOperator = (await listDevicePairing()).paired
+      .map(resolveEffectiveOperatorDeviceIdentity)
+      .find(
+        (device): device is EffectiveOperatorDeviceIdentity =>
+          device !== null &&
+          roleScopesAllow({
+            role: "operator",
+            requestedScopes: ["operator.pairing"],
+            allowedScopes: device.scopes,
+          }),
+      );
+    if (existingOperator) {
+      try {
+        controlUiDeviceAuthMigrationState = completeControlUiDeviceAuthMigration(
+          existingOperator.deviceId,
+          { env: process.env },
+        );
+      } catch (error) {
+        log.warn(
+          `failed to reconcile Control UI device-auth migration with existing operator: ${String(error)}`,
+        );
+      }
+    }
+  }
+  const controlUiDeviceAuthMigration = {
+    pending: controlUiDeviceAuthMigrationState?.status === "pending",
+  };
+  if (controlUiDeviceAuthMigration.pending) {
+    log.warn(
+      "Retired gateway.controlUi.dangerouslyDisableDeviceAuth config detected. Authenticated Control UI access remains available for pairing-only remediation; reopen the Control UI over HTTPS or localhost, then click Secure this browser.",
+    );
+  }
+  if (authBootstrap.generatedToken) {
+    log.warn(formatRuntimeGatewayAuthTokenWarning());
+  }
+  // prepareGatewayStartupConfig has already applied startupAuthOverride to cfgAtStart,
+  // so this warning follows the effective auth mode rather than dormant file config.
+  const trustedProxyDeviceAutoApprove = cfgAtStart.gateway?.auth?.trustedProxy?.deviceAutoApprove;
+  if (
+    cfgAtStart.gateway?.auth?.mode === "trusted-proxy" &&
+    trustedProxyDeviceAutoApprove?.enabled === true &&
+    trustedProxyDeviceAutoApprove.scopes?.some((scope) => scope.trim() === ADMIN_SCOPE)
+  ) {
+    log.warn(
+      "SECURITY WARNING: gateway.auth.trustedProxy.deviceAutoApprove.scopes includes operator.admin; every proxy-authenticated user can auto-approve a new browser device with full admin, and requests without scopes receive full admin automatically. Remove operator.admin to require manual approval until per-identity roles are available.",
+    );
+  }
+  const resolvedStartupAuthOverride = startupAuthOverride
+    ? (Object.fromEntries(
+        (
+          [
+            "mode",
+            "token",
+            "password",
+            "allowTailscale",
+            "rateLimit",
+            "trustedProxy",
+          ] as const satisfies readonly (keyof GatewayAuthConfig)[]
+        ).flatMap((key) => {
+          if (startupAuthOverride[key] === undefined) {
+            return [];
+          }
+          if ((key === "token" || key === "password") && isSecretRef(startupAuthOverride[key])) {
+            return [];
+          }
+          const resolvedValue = cfgAtStart.gateway?.auth?.[key];
+          return resolvedValue === undefined ? [] : [[key, structuredClone(resolvedValue)]];
+        }),
+      ) as GatewayAuthConfig)
+    : undefined;
+  const startupAuthSecretRefOverride = startupAuthOverride
+    ? {
+        ...(isSecretRef(startupAuthOverride.token)
+          ? { token: structuredClone(startupAuthOverride.token) }
+          : {}),
+        ...(isSecretRef(startupAuthOverride.password)
+          ? { password: structuredClone(startupAuthOverride.password) }
+          : {}),
+      }
+    : undefined;
+  const reloadAuthOverride = authBootstrap.generatedToken
+    ? mergeGatewayAuthConfig(resolvedStartupAuthOverride, { token: authBootstrap.generatedToken })
+    : resolvedStartupAuthOverride;
+  const diagnosticsEnabled = isDiagnosticsEnabled(cfgAtStart);
+  setDiagnosticsEnabledForProcess(diagnosticsEnabled);
+  if (diagnosticsEnabled) {
+    startDiagnosticHeartbeat(undefined, {
+      getConfig: getRuntimeConfig,
+      startupGraceMs: 60_000,
+    });
+  }
+  setGatewaySigusr1RestartPolicy({ allowExternal: isRestartEnabled(cfgAtStart) });
+  const activeTaskCount = { get: () => 0 };
+  setPreRestartDeferralCheck(
+    () =>
+      getTotalQueueSize() +
+      getTotalPendingReplies() +
+      getActiveEmbeddedRunCount() +
+      getActiveCronJobCount() +
+      getActiveBackgroundExecSessionCount() +
+      getActiveGatewayRootWorkCount({ excludeCurrent: true }) +
+      activeTaskCount.get(),
+  );
+  const seededControlUiAllowedOrigins = controlUiSeed.seededAllowedOrigins
+    ? cfgAtStart.gateway?.controlUi?.allowedOrigins
+    : undefined;
+  const applyFixedGatewayOverlays = (config: OpenClawConfig): OpenClawConfig => {
+    let runtimeConfig = config;
+    if (reloadAuthOverride || startupTailscaleOverride) {
+      runtimeConfig = {
+        ...runtimeConfig,
+        gateway: {
+          ...runtimeConfig.gateway,
+          ...(reloadAuthOverride
+            ? { auth: mergeGatewayAuthConfig(runtimeConfig.gateway?.auth, reloadAuthOverride) }
+            : {}),
+          ...(startupTailscaleOverride
+            ? {
+                tailscale: mergeGatewayTailscaleConfig(
+                  runtimeConfig.gateway?.tailscale,
+                  startupTailscaleOverride,
+                ),
+              }
+            : {}),
+        },
+      };
+    }
+    if (
+      seededControlUiAllowedOrigins &&
+      runtimeConfig.gateway?.controlUi?.allowedOrigins === undefined
+    ) {
+      runtimeConfig = {
+        ...runtimeConfig,
+        gateway: {
+          ...runtimeConfig.gateway,
+          controlUi: {
+            ...runtimeConfig.gateway?.controlUi,
+            allowedOrigins: seededControlUiAllowedOrigins,
+          },
+        },
+      };
+    }
+    return runtimeConfig;
+  };
+  const applyReloadableGatewayAuthRefs = (config: OpenClawConfig): OpenClawConfig => {
+    if (!startupAuthSecretRefOverride?.token && !startupAuthSecretRefOverride?.password) {
+      return config;
+    }
+    return {
+      ...config,
+      gateway: {
+        ...config.gateway,
+        auth: mergeGatewayAuthConfig(config.gateway?.auth, startupAuthSecretRefOverride),
+      },
+    };
+  };
+  const prepareReloadCandidate = (params: {
+    runtimeConfig: OpenClawConfig;
+    sourceConfig: OpenClawConfig;
+    previousSourceConfig?: OpenClawConfig;
+  }) => {
+    const previousSourceConfig =
+      params.previousSourceConfig ??
+      getRuntimeConfigSourceSnapshot() ??
+      startupLastGoodSnapshot.sourceConfig;
+    assertGatewayConfigEnvSelectionUnchanged(previousSourceConfig, params.sourceConfig);
+    const runtimeEnv = prepareConfigRuntimeEnv({
+      previousConfig: previousSourceConfig,
+      nextConfig: params.sourceConfig,
+    });
+    const metadata = startupConfigLoad.pluginMetadataSnapshot;
+    const pluginCandidate = minimalTestGateway
+      ? { runtimeConfig: params.runtimeConfig, compareConfig: params.sourceConfig }
+      : resolveGatewayReloadPluginActivationCandidate({
+          ...params,
+          env: runtimeEnv.env,
+          ...(metadata?.manifestRegistry ? { manifestRegistry: metadata.manifestRegistry } : {}),
+          discovery: metadata?.discovery,
+          ambientEnvTriggers,
+        });
+    const applyCandidateOverrides = captureConfigOverrideApplier();
+    const reapplyCompareOverlays = (config: OpenClawConfig): OpenClawConfig =>
+      applyCandidateOverrides(
+        mergeActivationSectionsIntoRuntimeConfig({
+          runtimeConfig: config,
+          activationConfig: pluginCandidate.compareConfig,
+        }),
+      );
+    const reapplyRuntimeOverlays = (config: OpenClawConfig): OpenClawConfig =>
+      applyFixedGatewayOverlays(applyReloadableGatewayAuthRefs(reapplyCompareOverlays(config)));
+    return {
+      runtimeConfig: reapplyRuntimeOverlays(params.runtimeConfig),
+      compareConfig: reapplyCompareOverlays(params.sourceConfig),
+      runtimeEnv,
+      reapplyRuntimeOverlays,
+      reapplyCompareOverlays,
+    };
+  };
+  // Keep the old startup-write suppression path intact for compatibility with
+  // callers that may still report a write, but startup itself no longer mutates config.
+  if (startupConfigLoad.wroteConfig || authBootstrap.persistedGeneratedToken) {
+    const startupSnapshot = await startupTrace.measure("config.final-snapshot", () =>
+      readConfigFileSnapshot(),
+    );
+    startupInternalWriteHash = startupSnapshot.hash ?? null;
+    startupLastGoodSnapshot = startupSnapshot;
+  }
+  setAppliedRuntimeConfigSnapshot(cfgAtStart, startupLastGoodSnapshot.sourceConfig);
+  initializePublishedConfigRuntimeEnv(startupLastGoodSnapshot.sourceConfig, {
+    ownedEnv: collectConfigRuntimeEnvOwnership(
+      startupLastGoodSnapshot.sourceConfig,
+      envBeforeStartupConfigLoad,
+      process.env,
+    ),
+    preserveExistingOwnership: true,
+  });
+  const workerEnvironmentStartup = minimalTestGateway
+    ? undefined
+    : await startupTrace.measure("worker-environments.store-import", async () => {
+        const workerModule = await loadWorkerEnvironmentStartupModule();
+        return await workerModule.loadGatewayWorkerEnvironmentStartupState();
+      });
+  const { prepareGatewayPluginBootstrap } = await loadStartupPluginsModule();
+  const pluginBootstrap = await startupTrace.measure("plugins.bootstrap", () =>
+    prepareGatewayPluginBootstrap({
+      cfgAtStart,
+      activationSourceConfig: startupActivationSourceConfig,
+      startupRuntimeConfig,
+      pluginMetadataSnapshot: startupConfigLoad.pluginMetadataSnapshot,
+      workerProviderIds: workerEnvironmentStartup?.durableProviderIds ?? [],
+      minimalTestGateway,
+      ambientEnvTriggers,
+      log,
+      loadRuntimePlugins: false,
+      loadSetupRuntimePlugins: true,
+    }),
+  );
+  const {
+    gatewayPluginConfigAtStart,
+    defaultWorkspaceDir,
+    deferredConfiguredChannelPluginIds,
+    startupPluginIds,
+    pluginLookUpTable,
+    baseMethods,
+    runtimePluginsLoaded,
+    ambientAutostartSuppressedChannelIds,
+  } = pluginBootstrap;
+  const coreGatewayMethodNames = listCoreGatewayMethodNames();
+  setCurrentPluginMetadataSnapshot(pluginLookUpTable, {
+    config: startupActivationSourceConfig,
+    compatibleConfigs: [startupRuntimeConfig, cfgAtStart, gatewayPluginConfigAtStart],
+    env: process.env,
+    workspaceDir: defaultWorkspaceDir,
+  });
+  if (pluginLookUpTable) {
+    const metrics = pluginLookUpTable.metrics;
+    startupTrace.detail("plugins.lookup-table", [
+      ["registrySnapshotMs", metrics.registrySnapshotMs],
+      ["manifestRegistryMs", metrics.manifestRegistryMs],
+      ["startupPlanMs", metrics.startupPlanMs],
+      ["ownerMapsMs", metrics.ownerMapsMs],
+      ["totalMs", metrics.totalMs],
+      ["indexPlugins", String(metrics.indexPluginCount)],
+      ["indexPluginCount", metrics.indexPluginCount],
+      ["manifestPlugins", String(metrics.manifestPluginCount)],
+      ["manifestPluginCount", metrics.manifestPluginCount],
+      ["startupPlugins", String(metrics.startupPluginCount)],
+      ["startupPluginCount", metrics.startupPluginCount],
+      ["deferredChannelPlugins", String(metrics.deferredChannelPluginCount)],
+      ["deferredChannelPluginCount", metrics.deferredChannelPluginCount],
+    ]);
+  }
+
+  return {
+    opts,
+    minimalTestGateway,
+    ambientEnvTriggers,
+    startupTrace,
+    loadStartupPluginsModule,
+    configSnapshot,
+    startupConfigLoad,
+    startupActivationSourceConfig,
+    startupRuntimeConfig,
+    cfgAtStart,
+    claimControlUiDeviceAuthMigration,
+    completeControlUiDeviceAuthMigration,
+    releaseControlUiDeviceAuthMigrationClaim,
+    controlUiDeviceAuthMigration,
+    resolvedStartupAuthOverride,
+    startupTailscaleOverride,
+    diagnosticsEnabled,
+    activeTaskCount,
+    applyFixedGatewayOverlays,
+    prepareReloadCandidate,
+    startupInternalWriteHash,
+    startupLastGoodSnapshot,
+    workerEnvironmentStartup,
+    pluginBootstrap,
+    gatewayPluginConfigAtStart,
+    defaultWorkspaceDir,
+    deferredConfiguredChannelPluginIds,
+    startupPluginIds,
+    pluginLookUpTable,
+    baseMethods,
+    runtimePluginsLoaded,
+    ambientAutostartSuppressedChannelIds,
+    coreGatewayMethodNames,
+    activateRuntimeSecrets,
+  };
+}
